# app/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Storefront API
  version: 1.0.0
  description: |
    REST API for the storefront backend — accounts, catalog, cart, checkout, orders,
    shipping addresses and an admin namespace.

    ## Base path

    Every API route is served under `/api/v1`, e.g. `POST /api/v1/auth/login`. The prose
    below leaves the prefix out. Only `GET /` (health check), `GET /docs` and
    `GET /openapi.yaml` sit at the server root.

    ## Response envelope

    Every route except `GET /` wraps its payload in the same envelope:

    ```json
    { "status": 200, "message": "Products fetched.", "data": [] }
    ```

    `status` repeats the HTTP status code and `data` is `null` on errors. An error adds a
    machine-readable `code`:

    ```json
    { "status": 401, "message": "Access token has expired.", "data": null, "code": "token_expired" }
    ```

    `code` is one of `invalid_request`, `bad_request`, `no_token`, `token_expired`,
    `token_invalid`, `invalid_credentials`, `forbidden`, `not_found`, `conflict`,
    `rate_limited` or `internal_error`.

    ## Pagination

    Lists that can grow — products, orders and the admin lists — take `limit`
    (1–100, default 50) and `offset` (default 0), and add a `meta` object next to `data`:

    ```json
    { "status": 200, "message": "Products fetched.", "data": [], "meta": { "limit": 50, "offset": 0, "total": 134 } }
    ```

    ## Authentication

...

    ```
    Authorization: Bearer <accessToken>
    ```

...

...

    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/logout-all:
    post:
      tags: [Auth]
      summary: Log out everywhere
      description: Deletes every refresh token of the signed-in account and clears the cookie.
      operationId: logoutAll
      responses:
        '200':
          description: Every session was revoked.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message: { type: string, example: All sessions revoked. }
                      data: { type: object, nullable: true, example: null }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/products/categories:
    get:
      tags: [Products]
      summary: Every category on sale
      description: The distinct categories of products that are on sale, sorted A–Z.
      operationId: listCategories
      responses:
        '200':
          description: The categories.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message: { type: string, example: Categories fetched. }
                      data:
                        type: array
                        items: { type: string }
                        example: [Clothing, Electronics, Furniture]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      summary: One of your orders
      description: Someone else's order answers `404`, so its existence is never revealed.
      operationId: getOwnOrder
      responses:
        '200':
          description: The order.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message: { type: string, example: Order fetched. }
                      data: { $ref: '#/components/schemas/Order' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFoundOrNotYours'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/cart:
    get:
      tags: [Cart]
      summary: Your cart
      description: Every row in the token user's cart, each with the product it points at.
      operationId: getCart
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message: { type: string, example: Cart fetched. }
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CartItem' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Cart]
      summary: Put something in the cart
      description: |
        Adds a product, or an option of it named by `typeId`. The same product and option
        twice adds up the quantities (capped at 999).

        The option, its price and the shop are read from the product; a price or option sent
        in the body is ignored. A product that does not exist or is archived answers `400`,
        and so does leaving `typeId` out for a product that has options.
      operationId: addCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CartItemInput' }
      responses:
        '201':
          description: The cart row.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      status: { type: integer, example: 201 }
                      message: { type: string, example: Item added to cart. }
                      data: { $ref: '#/components/schemas/CartItem' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Cart]
      summary: Empty the cart
      operationId: clearCart
      responses:
        '200':
          description: The cart is empty.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      message: { type: string, example: Cart cleared. }
                      data: { type: object, nullable: true, example: null }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/cart/checkout:
    post:
      tags: [Cart]
      summary: Buy the chosen cart rows
      description: |
        Turns the cart rows named by `cartItemIds` into one completed order, in a single
        transaction that locks those rows and the products they point at:

        - the quantity comes from the cart row, never from the request;
        - the price comes from the product, or from the chosen option, and is stored on the
          order line;
        - stock is checked and reduced on whichever row holds it: the option when one was
          chosen, otherwise the product;
        - the rows that were paid for leave the cart, and the rest stay.

        `addressId` must be one of the caller's own addresses; any other id answers `404`.
        It is copied onto the order, so editing or deleting that address later does not
        change where the order says it was sent.

        Anything that no longer adds up stops the whole checkout with `409` and changes
        nothing: an archived product, an option that is gone, or too little stock.
      operationId: checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartItemIds, addressId]
              properties:
                cartItemIds:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: integer, minimum: 1 }
                  example: [12, 13]
                addressId:
                  type: integer
                  minimum: 1
                  description: One of the caller's own addresses, to ship the order to.
                  example: 3
      responses:
        '201':
          description: The order and its lines.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      status: { type: integer, example: 201 }
                      message: { type: string, example: Checkout successful. }
                      data:
                        type: object
                        properties:
                          order: { $ref: '#/components/schemas/Order' }
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/OrderLine' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The cart no longer adds up, and nothing was charged.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                outOfStock:
                  summary: Not enough stock
                  value:
                    status: 409
                    message: Only 2 left of "Smart Fitness Tracker Watch"
                    data: null
                    code: conflict
                archived:
                  summary: A product was archived
                  value:
                    status: 409
                    message: One of those products is no longer on sale
                    data: null
                    code: conflict
                goneFromCart:
                  summary: A row is no longer in the cart
                  value:
                    status: 409
                    message: Some of those items are no longer in your cart
                    data: null
                    code: conflict
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The `accessToken` from `POST /auth/register`, `POST /auth/login` or
        `POST /auth/refresh`, sent as `Authorization: Bearer <accessToken>`.
        Expires after 15 minutes by default. The refresh token is not sent this way: it
        travels in an HttpOnly cookie.
...
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      description: Order id.
      schema: { type: integer, minimum: 1 }
      example: 1
...
  schemas:
    Envelope:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: integer
          description: Repeats the HTTP status code.
          example: 200
        message:
          type: string
          description: A sentence a client may show as-is.
          example: Products fetched.
        data:
          description: The payload, or `null` on an error.
          nullable: true
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          required: [code]
          properties:
            data: { type: object, nullable: true, example: null }
            code:
              type: string
              description: What went wrong, in a form a client can branch on.
              enum:
                - invalid_request
                - bad_request
                - no_token
                - token_expired
                - token_invalid
                - invalid_credentials
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - internal_error
...
    OrderStatus:
      type: string
      enum: [active, complete]
      example: complete
    AddressLabel:
      type: string
      enum: [home, work, other]
      example: home
    Money:
      type: string
      description: 'An amount with two decimals. `NUMERIC` in Postgres, so it travels as a string.'
      example: '129.99'
...
    ProductType:
      type: object
      description: One option of a product, with its own price and stock.
      properties:
        _id: { type: string, example: 69999b5d6decb17b3853fc1d }
        productId: { type: integer, example: 5002 }
        color: { type: string, example: Rose Gold }
        price: { type: number, format: float, example: 139.99 }
        stock: { type: integer, example: 40 }
        image: { type: string }
    Review:
      type: object
      properties:
        _id: { type: string }
        star: { type: number, minimum: 0, maximum: 5, example: 5 }
        comment: { type: string, example: Excellent fitness tracker! }
        userId: { type: string, example: user135 }
        userName: { type: string, example: Jennifer Martinez }
        date: { type: string, example: '2026-01-08T08:30:00.000Z' }
...
    Order:
      type: object
      properties:
        id: { type: integer, example: 7 }
        userId: { type: integer, example: 1 }
        status: { $ref: '#/components/schemas/OrderStatus' }
        createdAt: { type: string, format: date-time, example: '2026-09-22T10:42:05.000Z' }
        total:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: The sum of its lines, at the prices charged.
          example: '279.98'
        addressId:
          type: integer
          nullable: true
          description: The address it was shipped to, or null once that address is deleted.
          example: 3
        shippingAddress:
          allOf: [{ $ref: '#/components/schemas/ShippingAddress' }]
          nullable: true
          description: Copied onto the order when it was placed, so it never changes afterwards.
    ShippingAddress:
      type: object
      description: Where an order was sent, as it read at the moment of checkout.
      properties:
        fullName: { type: string, example: Jane Doe }
        phone: { type: string, nullable: true, example: '+44 20 7946 0958' }
        address: { type: string, example: 12 Market Street }
        city: { type: string, example: Manchester }
        label: { $ref: '#/components/schemas/AddressLabel' }
    OrderLine:
      type: object
      properties:
        id: { type: integer, example: 11 }
        orderId: { type: integer, example: 7 }
        productId: { type: integer, example: 1 }
        variantId:
          type: integer
          nullable: true
          description: The option row that was bought, or null once that option is retired.
          example: 42
        typeId:
          type: string
          nullable: true
          description: The name of the option that was bought, kept even if it is retired later.
          example: 69999b5d6decb17b3853fc1d
        quantity: { type: integer, example: 2 }
        unitPrice:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: What one of them cost when the order was placed.
          example: '139.99'
    CartItem:
      type: object
      description: A cart row, with the product it points at.
      properties:
        id: { type: integer, example: 12 }
        userId: { type: integer, example: 1 }
        productId: { type: integer, example: 1 }
        quantity: { type: integer, example: 2 }
        typeId: { type: string, nullable: true, example: 69999b5d6decb17b3853fc1d }
        selectedType:
          allOf: [{ $ref: '#/components/schemas/ProductType' }]
          nullable: true
          description: The chosen option, as the product defines it.
        shopId: { type: string, nullable: true, example: shop_001 }
        shopName: { type: string, nullable: true, example: TechZone Store }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        productName: { type: string, example: Smart Fitness Tracker Watch }
        productPrice: { $ref: '#/components/schemas/Money' }
        productCategory: { type: string, nullable: true, example: Electronics }
        productImage: { type: string, nullable: true }
        productDescription: { type: string, nullable: true }
        productPreviewImg:
          type: array
          items: { type: string }
        productTypes:
          type: array
          items: { $ref: '#/components/schemas/ProductType' }
        productReviews:
          type: array
          items: { $ref: '#/components/schemas/Review' }
        productOverallRating: { type: number, example: 4.7 }
        productStock: { type: integer, example: 145 }
        productIsActive: { type: boolean, example: true }
        productShopId: { type: string, nullable: true }
        productShopName: { type: string, nullable: true }
    CartItemInput:
      type: object
      required: [productId]
      description: |
        The option is named by id only. Its price, its stock and the shop are read from the
        product, so a price sent here is ignored.
      properties:
        productId: { type: integer, minimum: 1, example: 1 }
        quantity: { type: integer, minimum: 1, maximum: 999, default: 1, example: 2 }
        typeId:
          type: string
          nullable: true
          maxLength: 255
          example: 69999b5d6decb17b3853fc1d
...
  responses:
    BadRequest:
      description: The request does not add up. `message` names the field that was wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 400
            message: quantity must be a whole number between 1 and 999
            data: null
            code: invalid_request
    Unauthorized:
      description: |
        No token, an expired one, or one that does not check out. `code` tells a client
        whether to renew the session (`token_expired`) or sign in again.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            noToken:
              value: { status: 401, message: Access denied. No token provided., data: null, code: no_token }
            expired:
              value: { status: 401, message: Access token has expired., data: null, code: token_expired }
            invalid:
              value: { status: 401, message: Invalid token., data: null, code: token_invalid }
...
    NotFoundOrNotYours:
      description: |
        Nothing with that id, or it belongs to someone else — the two answer the same way, so
        the existence of another user's row is never revealed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 404
            message: order with id 999 not found
            data: null
            code: not_found
...
    InternalError:
      description: |
        Something went wrong on the server. The real error is logged with the request id from
        the `X-Request-Id` header; the answer stays deliberately vague.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: 500
            message: Internal Server Error
            data: null
            code: internal_error